Adjusted damage types.
I'm adding `DeathMagic` now to remind myself to make undead immune to it, so then they'll be immune to wands of death, which should be the only feature from IQ that will use this damage type.

I figure the bomb damage type is better named `Fire` because wands of flame will also use it.