Cleaned up how `EachTurn` works.
The old approach added each `EachTurn` instance to the current level's hook list in `EachTurn.__init__`, but object creation can occur when `G.level` is set to the wrong level, or the level we're going to put the object in doesn't even exist yet.

This change removes the ability of `EachTurn` objects to trigger while in the player's inventory, but was I ever going to use that?