Made scenery block monsters by default even with dainty monsters off.
Partly this is just a code simplification, but it has a gameplay effect, which is to make traps monster-impassible. This avoids the weirdness of monsters not triggering traps. And, it often forces monsters to make better tactical use of traps. For example, Death will hide behind a paralysis trap instead of charging over it.